feat: delete avatar controller

+++ src/modules/user/useCases/deleteUserAvatarUseCase/deleteUserAvatarUseCase.ts
@@ -1,4 +1,4 @@
-import { HttpException, HttpStatus } from '@nestjs/common';
+import { HttpException, HttpStatus, Injectable } from '@nestjs/common';
 import { BSON } from 'bson';
 import { UserRepository } from '../../repositories/userRepository';
 import * as fs from 'fs';
@@ -7,6 +7,7 @@ interface DeleteUserAvatarInterface {
   id: string;
 }
 
+@Injectable()
 export class DeleteUserAvatarUseCase {
   constructor(private userRepository: UserRepository) {}
 
@@ -18,6 +19,9 @@ export class DeleteUserAvatarUseCase {
 
     if (!currentUser.id)
       throw new HttpException('User not found', HttpStatus.NOT_FOUND);
+
+    if (!currentUser.avatar)
+      throw new HttpException(`User has no avatar`, HttpStatus.BAD_REQUEST);
 
     const avatarPath = `${process.cwd()}/uploads/${currentUser.avatar}`;
     await fs.promises.unlink(avatarPath);
